Hopefully final 1.15 build of v0.a11s5

+++ src/main/java/net/mcreator/wobr/procedures/BulletGroundHitProcedure.java
@@ -2,6 +2,8 @@ package net.mcreator.wobr.procedures;
 
 import net.minecraft.world.IWorld;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.particles.ParticleTypes;
 
 import net.mcreator.wobr.WobrModElements;
@@ -41,7 +43,9 @@ public class BulletGroundHitProcedure extends WobrModElements.ModElement {
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
 		IWorld world = (IWorld) dependencies.get("world");
 		world.addParticle(ParticleTypes.EXPLOSION, x, y, z, 0, 1, 0);
-		if (((world.getBlockState(new BlockPos((int) x, (int) y, (int) z))).getMaterial() == net.minecraft.block.material.Material.GLASS)) {
+		if ((((world.getBlockState(new BlockPos((int) x, (int) y, (int) z))).getMaterial() == net.minecraft.block.material.Material.GLASS)
+				&& (!(BlockTags.getCollection().getOrCreate(new ResourceLocation(("forge:non_breakable_glass").toLowerCase(java.util.Locale.ENGLISH)))
+						.contains((world.getBlockState(new BlockPos((int) x, (int) y, (int) z))).getBlock()))))) {
 			world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
 		}
 	}
